+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- feat: Transition from apiextensions.k8s.io/v1beta1 to
+  apiextensions.k8s.io/v1.
+
 # v0.5.1
 - [#1318](https://github.com/kubernetes-sigs/kubefed/pull/1318)
   chore: make certain cert-manager properties configurable.